backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0
psutil==5.9.6
python-decouple==3.8
whitenoise==6.6.0
django-extensions==3.2.3
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1